\name{Trio-class}
\docType{class}

% Class:
\alias{class:Trio}
\alias{Trio-class}
\alias{Trio}

\title{Trio Class}

\description{An object containing a character subject identifier \code{ID}, a character identifier of his/her \code{spouse} and a list of their \code{offspring}.}

\author{Samuel G. Younkin <syounkin@stat.wisc.edu>}
